Remove support for file names with spaces.
Target:addFiles does not support files with spaces in their names. I did
this to ease the things for 99.9% of meique future users, when you use:

t:addFiles("foo.cpp bar.cpp")

now has the same meaning of:

t:addFiles("foo.cpp", "bar.cpp")

but much less verbose.

As I never had seen a serious project with file names with spaces, I'll not implement
support for it, anyway it can easy be implemented in the future.