Adds __eval metamethod to Lua tables.
This modification to Lua language was needed to support scopes with "or"
operator. By default lua tables aways evals to true, after this patch
the __eval meta method will be called always Lua need to check if a value
is true or false. If the table doesn't have the __eval metamethod lua will
behave like it was used to, i.e. tables always evals to true.

After this modification you can use meique scopes in the following way:

(MACOSX or WIN32):target:addFiles("foo.cpp")

foo.cpp will be added to the target only if you are on a mac or on Windows 32bits.

If you want to use an and, the better way is to use:

WIN32:DEBUG:target:addFiles("foo.cpp")

which should be equivalent to:

(WIN32 and DEBUG):target:addFiles("foo.cpp")