Add __not metamethod to Lua tables.
This modification to Lua language was needed to support the negation of
scopes.

By default lua first transform a operand into a boolean and then applies the
not operator. For tables this implies in callign the (also new) __eval metamethod
and then applying the negation to the returned value interpreted as an boolean.
Now if the operand is a table or userdata it will try to use the metamethod __not,
if the metamethod was not found an error will be throw.

The behaviour for other types wasn't changed.